+++ lib/app.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:simbiotik_admin/core/configs/app_config.dart';
 
+import 'core/routers/app_router_configs.dart';
+
 Future<Widget> initializeApp(AppConfig config) async {
   WidgetsFlutterBinding.ensureInitialized();
   return MyApp(
@@ -10,19 +12,23 @@ Future<Widget> initializeApp(AppConfig config) async {
 
 class MyApp extends StatelessWidget {
   final AppConfig config;
-  const MyApp({
+  final AppRouterConfig _appRouter = AppRouterConfig();
+  MyApp({
     super.key,
     required this.config,
   });
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
+    return MaterialApp.router(
+      debugShowCheckedModeBanner: false,
       title: config.appName,
       theme: ThemeData(
         primarySwatch: Colors.blue,
       ),
-      home: const Scaffold(),
+      routerDelegate: _appRouter.router.routerDelegate,
+      routeInformationParser: _appRouter.router.routeInformationParser,
+      routeInformationProvider: _appRouter.router.routeInformationProvider,
     );
   }
 }