+++ CHANGELOG.md
@@ -17,3 +17,4 @@
 ## Added
 - Setting configuration flavor android and iOS
 - Setting environment
+- Login screen